+++ packages/node-core/src/utils/batch-size.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {getHeapStatistics} from 'v8';
+import {NodeConfig} from '../configure/NodeConfig';
 import {getLogger} from '../logger';
 
 const HIGH_THRESHOLD = 0.85;
@@ -9,14 +10,15 @@ const LOW_THRESHOLD = 0.6;
 
 const logger = getLogger('memory');
 
-export function checkMemoryUsage(batchSizeScale: number): number {
+export function checkMemoryUsage(batchSizeScale: number, nodeConfig: NodeConfig): number {
   const memoryData = getHeapStatistics();
   const ratio = memoryData.used_heap_size / memoryData.heap_size_limit;
   // TODO: fix where used
-  // if (argv.profiler) {
-  //   logger.info(`Heap Statistics: ${JSON.stringify(memoryData)}`);
-  //   logger.info(`Heap Usage: ${ratio}`);
-  // }
+  if (nodeConfig.profiler) {
+    logger.info(`Heap Statistics: ${JSON.stringify(memoryData)}`);
+    logger.info(`Heap Usage: ${ratio}`);
+  }
+
   let scale = batchSizeScale;
 
   if (ratio > HIGH_THRESHOLD) {
