+++ packages/node/src/utils/project.ts
@@ -24,7 +24,10 @@ import {
   SubstrateHandler,
   SubstrateHandlerKind,
 } from '@subql/common-substrate';
+import { NodeConfig, SubqueryRepo } from '@subql/node-core';
 import yaml from 'js-yaml';
+import Pino from 'pino';
+import { QueryTypes, Sequelize } from 'sequelize';
 import tar from 'tar';
 import { NodeVM, VMScript } from 'vm2';
 import { SubqlProjectDs } from '../configure/SubqueryProject';
@@ -298,4 +301,40 @@ export function loadChainTypesFromJs(
     );
   }
   return rawContent;
+}
+
+export async function getExistingProjectSchema(
+  nodeConfig: NodeConfig,
+  sequelize: Sequelize,
+  subqueryRepo: SubqueryRepo,
+  logger: Pino.Logger,
+): Promise<string> {
+  const DEFAULT_DB_SCHEMA = 'public';
+  let schema = nodeConfig.localMode ? DEFAULT_DB_SCHEMA : nodeConfig.dbSchema;
+
+  let schemas: string[];
+  try {
+    const result = await sequelize.query(
+      `SELECT schema_name FROM information_schema.schemata`,
+      {
+        type: QueryTypes.SELECT,
+      },
+    );
+    schemas = result.map((x: any) => x.schema_name) as [string];
+  } catch (err) {
+    logger.error(`Unable to fetch all schemas: ${err}`);
+    process.exit(1);
+  }
+  if (!schemas.includes(schema)) {
+    // fallback to subqueries table
+    const subqueryModel = await subqueryRepo.findOne({
+      where: { name: nodeConfig.subqueryName },
+    });
+    if (subqueryModel) {
+      schema = subqueryModel.dbSchema;
+    } else {
+      schema = undefined;
+    }
+  }
+  return schema;
 }
