+++ test/jest-setup.ts
@@ -1,2 +1,4 @@
 import 'reflect-metadata';
-import 'regenerator-runtime/runtime'
+import 'regenerator-runtime/runtime';
+// @ts-ignore
+import './jest.module';
