+++ test/jest.module.ts
@@ -1,0 +1,3 @@
+import {initLogger} from '@subql/node-core/logger';
+
+initLogger(true, 'json', 'info');
